[REFACTOR] 파일명 치환 CodeManager - ConnectionManager
